check task when is done
