+++ src/app/components/todolist/todolist.component.scss
@@ -7,7 +7,7 @@
   flex-direction: column;
   align-items: center;
   border-radius: 8px;
-  height: 360px;
+  height: 400px;
 
 
   .icons {
